Remove need for commit-msg hook
instead use rebase with reword helper to add missing commit-ids

fixes #267, #179

commit-id:fcc60e1a
