+++ .spr.yml
@@ -5,4 +5,5 @@ requireChecks: true
 requireApproval: true
 githubRemote: origin
 githubBranch: master
+remoteBranches: []
 mergeMethod: rebase
