--- cmd/commithook/main.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"os"
-
-	"github.com/ejoffe/spr/hook"
-)
-
-func main() {
-	hook.CommitHook(os.Args[1])
-}
