+++ cmd/reword/main.go
@@ -2,8 +2,11 @@ package main
 
 import (
 	"bufio"
+	"fmt"
 	"os"
 	"strings"
+
+	"github.com/google/uuid"
 )
 
 func main() {
@@ -31,10 +34,42 @@ func main() {
 		}
 		writefile.Close()
 	} else {
-		file, err := os.Open(filename)
-		check(err)
-		file.Close()
+		if shouldAppendCommitID(filename) {
+			appendCommitID(filename)
+		}
 	}
+}
+
+func shouldAppendCommitID(filename string) bool {
+	readfile, err := os.Open(filename)
+	check(err)
+	defer readfile.Close()
+
+	i := 0
+	nonEmptyCommitMessage := false
+	scanner := bufio.NewScanner(readfile)
+	for scanner.Scan() {
+		line := scanner.Text()
+		if line != "" && !strings.HasPrefix(line, "#") {
+			nonEmptyCommitMessage = true
+		}
+		if strings.HasPrefix(line, "commit-id:") {
+			return false
+		}
+		i++
+	}
+	check(scanner.Err())
+	return nonEmptyCommitMessage
+}
+
+func appendCommitID(filename string) {
+	appendfile, err := os.OpenFile(filename, os.O_APPEND|os.O_WRONLY, 0666)
+	check(err)
+	defer appendfile.Close()
+
+	commitID := uuid.New()
+	appendfile.WriteString("\n")
+	appendfile.WriteString(fmt.Sprintf("commit-id:%s\n", commitID.String()[:8]))
 }
 
 func check(err error) {
